Move legacy-messages.ts to /query-server
