+++ extensions/ql-vscode/src/run-queries-shared.ts
@@ -1,5 +1,5 @@
 import * as messages from "./pure/messages-shared";
-import * as legacyMessages from "./pure/legacy-messages";
+import * as legacyMessages from "./query-server/legacy-messages";
 import { DatabaseInfo, QueryMetadata } from "./common/interface-types";
 import { join, parse, dirname, basename } from "path";
 import {
