+++ extensions/ql-vscode/test/vscode-tests/cli-integration/legacy-query.test.ts
@@ -3,7 +3,7 @@ import { join, basename } from "path";
 import { dirSync } from "tmp";
 import { pathToFileURL } from "url";
 import { CancellationTokenSource } from "vscode-jsonrpc";
-import * as messages from "../../../src/pure/legacy-messages";
+import * as messages from "../../../src/query-server/legacy-messages";
 import * as qsClient from "../../../src/query-server/legacy/query-server-client";
 import * as cli from "../../../src/codeql-cli/cli";
 import { CellValue } from "../../../src/common/bqrs-cli-types";
