+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-history/store/query-history-store.test.ts
@@ -12,7 +12,7 @@ import { QueryWithResults } from "../../../../../src/run-queries-shared";
 import { DatabaseInfo } from "../../../../../src/common/interface-types";
 import { CancellationTokenSource, Uri } from "vscode";
 import { tmpDir } from "../../../../../src/tmp-dir";
-import { QueryResultType } from "../../../../../src/pure/legacy-messages";
+import { QueryResultType } from "../../../../../src/query-server/legacy-messages";
 import { QueryInProgress } from "../../../../../src/query-server/legacy";
 import { VariantAnalysisHistoryItem } from "../../../../../src/query-history/variant-analysis-history-item";
 import { QueryHistoryInfo } from "../../../../../src/query-history/query-history-info";
